requirements.txt
Flask==3.0.2
Flask-Login==0.6.3
Flask-WTF==1.2.1
email-validator==2.1.0.post1
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.2
redis==5.0.4
rq==1.16.2
python-dotenv==1.0.1
boto3==1.34.144
sendgrid==6.11.0
openai==1.35.7
Flask-SQLAlchemy==3.1.1
gunicorn==21.2.0